Modified how the progress bar is being added
As we discussed during the call, the logic is that the entire `status-bar`, containing the progress bar, is created in `progressBar.js` and not in `questionView.js`. Then, once `initQuestionPage()` is called on `questionPage.js`, it checks whether the progress bar exists or not. If it doesn't, then it calls `createProgressBarElement()` which returns the `status-bar` container, which is then prepended to the body. I made a few minor style changes and updated the score counter for this new logic.